+++ src/eva2/tools/math/Jama/Matrix.java
@@ -1353,7 +1353,7 @@ public class Matrix implements Cloneable, Serializable {
         int n = v.size();  // Now we've got the number of columns!
         double row[] = new double[n];
         for (int j = 0; j < n; j++) {
-            row[j] = ((Double) v.elementAt(j)).doubleValue();
+            row[j] = (Double) v.elementAt(j);
         }
         v.removeAllElements();
         v.addElement(row);  // Start storing rows instead of columns.
@@ -1366,7 +1366,7 @@ public class Matrix implements Cloneable, Serializable {
                     throw new java.io.IOException
                             ("Row " + v.size() + " is too long.");
                 }
-                row[j++] = Double.valueOf(tokenizer.sval).doubleValue();
+                row[j++] = Double.valueOf(tokenizer.sval);
             } while (tokenizer.nextToken() == StreamTokenizer.TT_WORD);
             if (j < n) {
                 throw new java.io.IOException
